BaseReqView is added to reuse the req panel and HomeView and YourReqsView is added as the child of BaseReqView. few methods are added to the Query class to make the query easier and keep the code clean
