+++ DisasterDonationReliefManagementSystem/Entities/Delivery.cs
@@ -8,10 +8,20 @@ namespace DisasterDonationReliefManagementSystem.Entities
     {
         public int DeliveryID { get; set; }
         public int DonationID { get; set; }
-        public int UserID { get; set; }
+        public int VolunteerID { get; set; }
         public string PickupLocation { get; set; }
         public string DeliveryLocation { get; set; }
         public string DeliveryStatus { get; set; }
-    }
 
+        public Delivery(int deliveryID, int donationID, int volID, string pickupLocation, string deliveryLocation, string deliveryStatus)
+        {
+            DeliveryID = deliveryID;
+            DonationID = donationID;
+            VolunteerID = volID;
+            PickupLocation = pickupLocation;
+            DeliveryLocation = deliveryLocation;
+            DeliveryStatus = deliveryStatus;
+
+        }
+    }
 }
